Removed Global Variables

+++ uninstall-nfs-server.py
@@ -2,36 +2,34 @@
 import sys
 import os
 import shutil
-## Uninstall function
-def uninstall_server():  
-    os.system('sudo apt-get remove -y nfs-kernel-server')
-    os.remove("/etc/exports")
-    shutil.copy('/etc/exports.bak', '/etc/exports')
-    os.remove("/etc/exports.bak")
 
-## Function to validate yes no input.
+dry_run = True  # Set to true to not make changes for testing
+
 def yes_or_no(question):
-    global reply
-    answer = input(question + "(yes/no): ").lower().strip()
-    print("")
-    while not(answer == "y" or answer == "yes" or \
-    answer == "n" or answer == "no"):
-        print("Input yes or no")
-        answer = input(question + "(yes/no):").lower().strip()
-        print("")
+    '''Function to validate yes no input.'''
+    answer = None
+    while answer not in ["yes", "y", "no", "n"]:
+        if answer is not None:
+            print("Input yes or no")
+        answer = input(question + "(yes/no): ").lower().strip()
     if answer[0] == "y":
         return True
-        reply = '1'
     if answer[0] == "n":
-        reply = '0'
-    else:
         return False
 
-reply = ' '
-question = yes_or_no("Do you want to uninstall the  NFS server?  ")
-if reply == '0':
-   print("Exiting Setup")
-   sys.exit()
+def uninstall_server(): 
+     '''Function to uninstall nfs server.'''
+     if dry_run:
+        print("Uninstalling nfs-kernel-server")
+     else:
+        print("Uninstalling nfs-kernel-server")
+        os.system('sudo apt-get remove -y nfs-kernel-server')
+        os.remove("/etc/exports")
+        shutil.copy('/etc/exports.bak', '/etc/exports')
+        os.remove("/etc/exports.bak")
 
-print("Uninstalling Server")
+if not yes_or_no("Do you want to setup an NFS server?"):
+    print("Exiting Setup")
+    sys.exit()
+
 uninstall_server()
